Collect power samples from MCP39F501. Send those samples over MQTT to cloud server.

+++ cloud.h
@@ -21,7 +21,7 @@ typedef struct telemetryData_s
 {
 	uint8_t data[32];
 	uint8_t clientPhyAddr[8];
-	int len;
+	uint8_t len;
 } TelemetryData;
 
 extern QueueHandle_t xMqttQueue;
